[central] JobController -> JobPoller
