+++ pkg/central/server/gommon.yml
@@ -5,5 +5,5 @@ loggers:
     receiver: "srv"
   - struct: "*Manager"
     receiver: "mgr"
-  - struct: "*JobController"
+  - struct: "*JobPoller"
     receiver: "j"
