+++ pkg/central/server/job.go
@@ -1,45 +1,61 @@
 package server
 
 import (
+	"bytes"
 	"context"
+	"fmt"
+	"sync/atomic"
 	"time"
 
+	dconfig "github.com/dyweb/gommon/config"
+	"github.com/dyweb/gommon/errors"
 	dlog "github.com/dyweb/gommon/log"
 
 	pb "github.com/benchhub/benchhub/pkg/bhpb"
-
-	"github.com/dyweb/gommon/errors"
 )
 
-// Deprecated
-type JobController struct {
+// JobPoller get job for store and create job managers to run them
+type JobPoller struct {
 	registry *Registry
 	log      *dlog.Logger
 }
 
-type AssignResult struct {
-	Spec pb.NodeSelectorSpec
-	Node pb.Node
-}
-
-// Deprecated
-func NewJobController(r *Registry) (*JobController, error) {
-	j := &JobController{
+func NewJobPoller(r *Registry) (*JobPoller, error) {
+	j := &JobPoller{
 		registry: r,
 	}
 	dlog.NewStructLogger(log, j)
 	return j, nil
 }
 
-func (j *JobController) RunWithContext(ctx context.Context) error {
+func (srv *GrpcServer) SubmitJob(ctx context.Context, req *pb.SubmitJobReq) (*pb.SubmitJobRes, error) {
+	var job pb.JobSpec
+	if err := dconfig.LoadYAMLDirectFromStrict(bytes.NewReader([]byte(req.Spec)), &job); err != nil {
+		return nil, errors.Wrap(err, "can't parse YAML job spec")
+	}
+	// TODO: implement the validate logic
+	//if err := job.Validate(); err != nil {
+	//	return nil, errors.Wrap(err, "invalid job spec")
+	//}
+	// TODO: wrap this in store, store should return an id for job ...
+	// FIXME: we are just using project name + a global counter ...
+	atomic.AddInt64(&srv.c, 1)
+	id := fmt.Sprintf("%s-%d", job.Name, atomic.LoadInt64(&srv.c))
+	srv.log.Infof("got job %s id %s", job.Name, id)
+	if err := srv.meta.AddJobSpec(id, job); err != nil {
+		return nil, errors.Wrap(err, "can't add job spec to store")
+	}
+	return &pb.SubmitJobRes{Id: id}, nil
+}
+
+func (j *JobPoller) RunWithContext(ctx context.Context) error {
 	j.log.Info("start job controller")
 	meta := j.registry.Meta
 	for {
 		select {
 		case <-ctx.Done():
-			// TODO: should we return nil or return context error?
-			// TODO: we should tell all the agent to abort job since central is shut down?
-			j.log.Infof("job controller stop due to context finished, its error is %v", ctx.Err())
+			// TODO: tell all managers to stop?
+			j.log.Infof("job poller stop due to context finished, its error is %v", ctx.Err())
 			return nil
 		default:
 			job, empty, err := meta.GetPendingJob()
@@ -48,88 +64,12 @@ func (j *JobController) RunWithContext(ctx context.Context) error {
 			} else if err != nil {
 				log.Warnf("failed to get pending job %v", err)
 			} else {
-				// TODO: spec should contains id ...
-				_, err := meta.ListNodes()
-				if err != nil {
-					log.Warnf("can't list nodes %s", err.Error())
-				}
-				// TODO: acquire node resource based on node selector
-				//results, err := j.AcquireNodes(nodes, job.NodeAssignments)
-				//if err != nil {
-				//	log.Warnf("can't acquire nodes %s", err.Error())
-				//}
-				//for _, r := range results {
-				//	// TODO: print it or? ...
-				//	log.Infof("result is %v", r)
-				//}
-				log.Infof("TODO: process job %s", job.Name)
+				// TODO: get nodes and assign jobs
+				// TODO: put back if failed to schedule
+				log.Infof("TODO: deal with job %s", job.Id)
 			}
 			// TODO: poll duration should be configurable
 			time.Sleep(1 * time.Second)
 		}
 	}
-}
-
-func (j *JobController) AcquireNodes(nodes []pb.Node, specs []pb.NodeSelectorSpec) ([]AssignResult, error) {
-	if len(nodes) == 0 {
-		return nil, errors.New("0 agent no node to acquire")
-	}
-	if len(nodes) < len(specs) {
-		j.log.Warnf("only %d agents but want %d nodes", len(nodes), len(specs))
-	}
-	used := make([]int, len(nodes))
-	acquired := make([]int, len(specs))
-	// first loop, don't reuse any node
-
-	for i, s := range specs {
-		for j, node := range nodes {
-			if used[j] > 0 {
-				continue
-			}
-			if (s.Role == pb.Role_DATABASE && node.Info.Role == pb.Role_DATABASE) ||
-				(s.Role == pb.Role_LOADER && node.Info.Role == pb.Role_LOADER) {
-				// NOTE: +1 so we can check if the spec has acquired node with > 0
-				acquired[i] = j + 1
-				used[j]++
-				break
-			}
-			if node.Info.Role == pb.Role_ANY {
-				acquired[i] = j + 1
-				used[j]++
-				break
-			}
-		}
-	}
-	// second loop,
-	// TODO: allow multiple workload on node, but never assign database with loader on same node
-	for maxUse := 0; maxUse < 3; maxUse++ {
-		for i := range specs {
-			if acquired[i] > 0 {
-				continue
-			}
-			for j := range nodes {
-				if used[j] > maxUse {
-					continue
-				}
-				acquired[i] = j + 1
-				used[j]++
-				break
-			}
-		}
-	}
-	// all spec assigned to nodes?
-	merr := errors.NewMultiErr()
-	results := make([]AssignResult, 0, len(specs))
-	for i, spc := range specs {
-		if acquired[i] > 0 {
-			j.log.Infof("plan: assign %s %s to node %s", spc.Name, spc.Role, nodes[acquired[i]-1].Id)
-			results = append(results, AssignResult{
-				Spec: spc,
-				Node: nodes[acquired[i]-1],
-			})
-		} else {
-			merr.Append(errors.Errorf("spec %s %s has no node", spc.Name, spc.Role))
-		}
-	}
-	return results, merr.ErrorOrNil()
 }
